generate EH info for asm statements (PR93981)
The following patch extends the generation of exception handling
information, so that it is possible to catch exceptions thrown from
asm statements, when -fnon-call-exceptions is enabled.  Parts of the
gcc code already suggested this should be possible, but it was never
fully implemented.  Both volatile and non-volatile asms, including
asm goto, are now considered potentially throwing.

A new '-' constraint modifier is introduced, that allows users to
specify which asm outputs must be clobbered in case an exception is
thrown.  This also applies to the jump path for asm goto.

Three new test cases are added.  The target-dependent test should pass
on platforms where throwing from a signal handler is allowed.  The only
platform I am aware of where that is the case is *-linux-gnu, so it is
set to XFAIL on all others.

gcc/
2020-12-01  Jan W. Jagersma  <jwjagersma@gmail.com>

	PR inline-asm/93981
	* cfgexpand.c (expand_asm_stmt): Expand throwing asms into jump
	insns.  Emit move insns across edges.
	* gimplify.c (gimplify_asm_expr): Assign '-' constrained
	outputs via a temporary.  Warn for other constraint modifiers
	if a copy must be made.
	* stmt.c (parse_output_constraint): Accept '-' modifier.
	(parse_input_constraint): Don't accept '-' modifier.
	* tree-cfg.c (make_edges_bb): Make EH edges for GIMPLE_ASM.
	* tree-eh.c (lower_eh_constructs_2): Add case for GIMPLE_ASM.
	(tree_could_trap_p): Return true for case ASM_EXPR.
	(stmt_could_throw_p): Likewise, for GIMPLE_ASM.
	* rtlanal.c (may_trap_p_1): Likewise, for ASM_OPERANDS.
	* doc/extend.texi: Document the effects of constraint modifiers
	on output semantics for throwing/jumping asms.
	* doc/md.texi: Document new '-' constraint modifier.

gcc/testsuite/
2020-12-01  Jan W. Jagersma  <jwjagersma@gmail.com>

	PR inline-asm/93981
	* g++.target/i386/pr93981.C: New test.
	* g++.dg/eh/pr93981-1.C: New test.
	* g++.dg/eh/pr93981-2.C: New test.